+++ railgun/manager/run.go
@@ -120,6 +120,12 @@ func Run(ctx context.Context, c *config.Config) error {
 	}
 
 	// determine the proxy synchronization strategy
+	if c.ProxySyncSeconds < proxy.DefaultSyncSeconds {
+		setupLog.Info(fmt.Sprintf("WARNING: --proxy-sync-seconds is configured for %fs, in DBLESS mode this may result in"+
+			" problems of inconsistency in the proxy state. For DBLESS mode %fs+ is recommended (3s is the default).",
+			c.ProxySyncSeconds, proxy.DefaultSyncSeconds,
+		))
+	}
 	syncTickDuration, err := time.ParseDuration(fmt.Sprintf("%gs", c.ProxySyncSeconds))
 	if err != nil {
 		setupLog.Error(err, "%s is not a valid number of seconds to stagger the proxy server synchronization")